+++ apps/server/instamint-business/src/routes/admin/users/prepareAdminUsersActionsRoutes.ts
@@ -11,10 +11,11 @@ import {
   adminMessages,
   authMessages,
   globalsMessages,
+  redisKeys,
   usersMessages,
 } from "@/def"
 import { throwInternalError } from "@/utils/errors/throwInternalError"
-import { nowDate } from "@/utils/helpers/times"
+import { nowDate, sixMonthsDate } from "@/utils/helpers/times"
 
 const prepareAdminUsersActionsRoutes: ApiRoutes = ({ app, db, redis }) => {
   const adminUsersActions = new Hono()
@@ -92,6 +93,42 @@ const prepareAdminUsersActionsRoutes: ApiRoutes = ({ app, db, redis }) => {
         {
           message: adminMessages.userReactivatedSuccessfully.message,
         },
+        SC.success.OK
+      )
+    }
+  )
+
+  adminUsersActions.delete(
+    "/:id/delete",
+    zValidator("param", userIdAdminAction),
+    async (c: Context): Promise<Response> => {
+      const { id } = c.req.param() as UserIdAdminAction
+
+      const user = await UserModel.query().findById(id)
+
+      if (!user) {
+        return c.json(authMessages.userNotFound, SC.errors.NOT_FOUND)
+      }
+
+      if (!user.active || user.deletionDate !== null) {
+        return c.json(
+          usersMessages.accountAlreadyDeactivated,
+          SC.errors.BAD_REQUEST
+        )
+      }
+
+      await UserModel.query().where({ email: user.email }).update({
+        active: false,
+        deactivationDate: nowDate,
+        deletionDate: sixMonthsDate,
+      })
+
+      const sessionKey = redisKeys.auth.authSession(user.email)
+
+      await redis.del(sessionKey)
+
+      return c.json(
+        { message: usersMessages.deletedAccount.message },
         SC.success.OK
       )
     }